server/python-services/requirements.txt
monai==1.3.0
numpy==1.24.3
Pillow==10.0.0
scikit-image==0.21.0
vtk==9.2.6
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1 